Update to just create libc-client package
Case EA-4087: This commit takes out a lot of the extra libraries
that were created by EPEL, and just emits a libc-client.  It's
basically a skinnified version of EPEL's package.

The EPEL package originally created a uw-imap (and friends)
package, as well as a libc-client package.  By releasing the
uw-imap package, that can easily confuse people about why
the package exists, and are we actually using it.