Updated libc-client package for SCL compatibility
Case EA-4091: This updates the SPEC file with SCL conditional
macros.  If someone were to build this package without defining
the 'scl' namespace macro, then it should build in the typical
/usr/lib64 path.

With these alterations, if a user defines the 'scl' namespace
macro, then it builds the package within the corresponding
namespace.