Fix Makefile when attempting to build multiple SCL packages
Case EA-4094: Moves the '@' sign to the beginning of the make
line.  Placing it anywhere else was preventing it from building
more than 1 at a time.

That said, it would be nice to have a 'TODO' will allow the dev
to build a single SCL package, instead of all of them every time.